Rename package containing FocusArea.java and Major.java
Rename to "course" to better reflect the purpose of the classes
