+++ src/main/java/seedu/planner/model/ModelManager.java
@@ -14,8 +14,8 @@ import javafx.collections.ObservableList;
 import seedu.planner.commons.core.ComponentManager;
 import seedu.planner.commons.core.LogsCenter;
 import seedu.planner.commons.events.model.ModulePlannerChangedEvent;
-import seedu.planner.model.enumeration.FocusArea;
-import seedu.planner.model.enumeration.Major;
+import seedu.planner.model.course.FocusArea;
+import seedu.planner.model.course.Major;
 import seedu.planner.model.module.Module;
 import seedu.planner.model.module.ModuleInfo;
 import seedu.planner.model.module.ModuleType;
